refactor ProtocolDetails
