+++ packages/kit/src/views/Staking/components/ProtocolDetails/FaqSection.tsx
@@ -1,0 +1,109 @@
+import { useIntl } from 'react-intl';
+
+import {
+  Accordion,
+  Icon,
+  SizableText,
+  Stack,
+  YStack,
+} from '@onekeyhq/components';
+import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
+import { usePromiseResult } from '@onekeyhq/kit/src/hooks/usePromiseResult';
+import { ETranslations } from '@onekeyhq/shared/src/locale';
+import type { IStakeProtocolDetails } from '@onekeyhq/shared/types/staking';
+
+type ISolutions = {
+  question: string;
+  answer: string;
+}[];
+
+function FaqInfo({ solutions }: { solutions: ISolutions }) {
+  const intl = useIntl();
+  return (
+    <YStack gap="$6">
+      <SizableText size="$headingLg">
+        {intl.formatMessage({ id: ETranslations.global_faqs })}
+      </SizableText>
+      <YStack>
+        <Accordion type="multiple" gap="$2">
+          {solutions.map(({ question, answer }, index) => (
+            <Accordion.Item value={String(index)} key={String(index)}>
+              <Accordion.Trigger
+                unstyled
+                flexDirection="row"
+                alignItems="center"
+                borderWidth={0}
+                bg="$transparent"
+                px="$2"
+                py="$1"
+                mx="$-2"
+                my="$-1"
+                hoverStyle={{
+                  bg: '$bgHover',
+                }}
+                pressStyle={{
+                  bg: '$bgActive',
+                }}
+                borderRadius="$2"
+              >
+                {({ open }: { open: boolean }) => (
+                  <>
+                    <SizableText
+                      textAlign="left"
+                      flex={1}
+                      size="$bodyLgMedium"
+                      color={open ? '$text' : '$textSubdued'}
+                    >
+                      {question}
+                    </SizableText>
+                    <Stack animation="quick" rotate={open ? '180deg' : '0deg'}>
+                      <Icon
+                        name="ChevronDownSmallOutline"
+                        color={open ? '$iconActive' : '$iconSubdued'}
+                        size="$5"
+                      />
+                    </Stack>
+                  </>
+                )}
+              </Accordion.Trigger>
+              <Accordion.HeightAnimator animation="quick">
+                <Accordion.Content
+                  unstyled
+                  pt="$2"
+                  pb="$5"
+                  animation="quick"
+                  enterStyle={{ opacity: 0 }}
+                  exitStyle={{ opacity: 0 }}
+                >
+                  <SizableText size="$bodyMd">{answer}</SizableText>
+                </Accordion.Content>
+              </Accordion.HeightAnimator>
+            </Accordion.Item>
+          ))}
+        </Accordion>
+      </YStack>
+    </YStack>
+  );
+}
+
+export const FaqSection = ({
+  details,
+}: {
+  details?: IStakeProtocolDetails;
+}) => {
+  const { result: solutions } = usePromiseResult(
+    async () =>
+      details
+        ? backgroundApiProxy.serviceStaking.getFAQList({
+            symbol: details.token.info.symbol,
+            provider: details.provider.name,
+          })
+        : Promise.resolve([]),
+    [details],
+    {
+      initResult: [],
+    },
+  );
+  if (solutions.length === 0) return null;
+  return <FaqInfo solutions={solutions} />;
+};
